requirements.txt
tweepy
python-dotenv
requests
openai
Pillow
